+++ app/src/main/java/com/juanantbuit/weatherproject/usecases/GetCityInfoUseCase.kt
@@ -6,6 +6,6 @@ import com.juanantbuit.weatherproject.domain.models.CityInfoModel
 class GetCityInfoUseCase {
 
     private val repository = CityInfoRepository()
-    suspend fun getCityInfo(latitude: Double?, longitude: Double?): CityInfoModel = repository.getCityInfo(latitude, longitude)
+    suspend fun getCityInfo(latitude: Float?, longitude: Float?): CityInfoModel = repository.getCityInfo(latitude, longitude)
 
 }
