+++ app/src/main/java/com/juanantbuit/weatherproject/usecases/GetForecastResponseUseCase.kt
@@ -6,6 +6,6 @@ import com.juanantbuit.weatherproject.domain.models.ForecastResponseModel
 class GetForecastResponseUseCase {
 
     private val repository = ForecastResponseRepository()
-    suspend fun getForecastResponse(latitude: Double?, longitude: Double?): ForecastResponseModel = repository.getForecastResponse(latitude, longitude)
+    suspend fun getForecastResponse(latitude: Float?, longitude: Float?): ForecastResponseModel = repository.getForecastResponse(latitude, longitude)
 
 }
